requirements.txt
Django==2.2.5
djangorestframework==3.10.3
django-rest-swagger==2.2.0
